save and create file to send mail

+++ database/migrations/2023_07_18_080128_create_mm_to_sg_items_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('mm_to_sg_items', function (Blueprint $table) {
             $table->id();
             $table->string('sender_name');
-            $table->integer('sender_phone');
+            $table->bigInteger('sender_phone');
             $table->string('sender_address');
             $table->tinyInteger('transport')->comment('1: Sea Cargo 2: Air Cargo');
             $table->tinyInteger('storage_type')->comment('1: Room Temperature 2: In Normal Fridge 3: In Freezer');
@@ -25,7 +25,7 @@ return new class extends Migration
             $table->string('invoice_no');
             $table->tinyInteger('payment_type')->comment('1: SG Pay 2: MM Pay');
             $table->string('receiver_name');
-            $table->integer('receiver_phone');
+            $table->bigInteger('receiver_phone');
             $table->string('receiver_address');
             $table->string('receiver_postal_code')->nullable();
             $table->longText('additional_instruction')->nullable();
